Move 'CardFooter' subcomponent to `card-footer` subfolder, refactor hook out
