+++ packages/components/src/card/card-footer/README.md
@@ -1,0 +1,7 @@
+# CardFooter
+
+<div class="callout callout-alert">
+This feature is still experimental. “Experimental” means this is an early implementation subject to drastic and breaking changes.
+</div>
+
+TBD
