+++ packages/components/src/card/card-footer/index.js
@@ -1,0 +1,2 @@
+export { default } from './component';
+export { useCardFooter } from './hook';
