+++ packages/components/src/card/index.js
@@ -1,5 +1,5 @@
 export { default as Card, useCard } from './card';
+export { default as CardFooter, useCardFooter } from './card-footer';
 
 export { default as CardBody } from './body';
-export { default as CardFooter } from './footer';
 export { default as CardHeader } from './header';
